black applied

+++ pipelines/mlserver-final/router-with-queue/client-async-grpc.py
@@ -7,6 +7,7 @@ import mlserver.grpc.dataplane_pb2_grpc as dataplane
 import mlserver.grpc.converters as converters
 from mlserver.codecs.string import StringRequestCodec
 from mlserver.codecs.string import StringRequestCodec
+
 pp = PrettyPrinter(indent=4)
 from datasets import load_dataset
 import mlserver.types as types
@@ -21,9 +22,10 @@ async def send_requests(ch, payload, metadata):
         payload, model_name=model, model_version=None
     )
     response = await grpc_stub.ModelInfer(
-        request=inference_request_g,
-        metadata=metadata)
+        request=inference_request_g, metadata=metadata
+    )
     return response
+
 
 # single node mlserver
 # endpoint = "localhost:8081"
@@ -33,20 +35,19 @@ async def send_requests(ch, payload, metadata):
 
 # single node seldon+mlserver
 endpoint = "localhost:32000"
-deployment_name = 'router'
-model = 'router'
+deployment_name = "router"
+model = "router"
 namespace = "default"
 metadata = [("seldon", deployment_name), ("namespace", namespace)]
 
 batch_test = 5
 ds = load_dataset(
-    "hf-internal-testing/librispeech_asr_demo",
-    "clean",
-    split="validation")
+    "hf-internal-testing/librispeech_asr_demo", "clean", split="validation"
+)
 
 input_data = ds[0]["audio"]["array"]
 data_shape = [len(input_data)]
-custom_parameters = {'custom_2': 'test_2'}
+custom_parameters = {"custom_2": "test_2"}
 payload = types.InferenceRequest(
     inputs=[
         types.RequestInput(
@@ -55,9 +56,8 @@ payload = types.InferenceRequest(
             datatype="BYTES",
             data=[input_data.tobytes()],
             parameters=types.Parameters(
-                dtype='f4',
-                datashape=str(data_shape),
-                **custom_parameters),
+                dtype="f4", datashape=str(data_shape), **custom_parameters
+            ),
         )
     ]
 )
@@ -65,7 +65,9 @@ payload = types.InferenceRequest(
 
 async def main():
     async with grpc.aio.insecure_channel(endpoint) as ch:
-        responses = await asyncio.gather(*[send_requests(ch, payload, metadata) for _ in range(10)])
+        responses = await asyncio.gather(
+            *[send_requests(ch, payload, metadata) for _ in range(10)]
+        )
 
     # inference_responses = list(map(
     #     lambda response: ModelInferResponseConverter.to_types(response), responses))
@@ -76,5 +78,6 @@ async def main():
     #     lambda raw_json: json.loads(raw_json[0]), raw_jsons))
 
     pp.pprint(responses)
+
 
 asyncio.run(main())
